Refactor: Split models, views into multiple apps
