+++ WellPath/WellPath/settings.py
@@ -37,7 +37,9 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'goals'
+    'goals',
+    'social',
+    'taxonomy',
 ]
 
 MIDDLEWARE = [
