+++ WellPath/goals/urls.py
@@ -17,9 +17,7 @@ urlpatterns = [
     path('add_progress/', views.add_progress, name='add_progress'),  
     path('history/<int:goal_id>/', views.progress_history, name='progress_history'),
     path("api/goals", views.goals_api, name="goals_api"),
-    path("goals/<int:goal_id>/like/", views.like_goal, name="like_goal"),
-    path("goals/<int:goal_id>/comments/", views.comment_goal, name="comment_goal"),
-    path('category/<slug:category_slug>/', views.category, name='category')
+
 
 
 ]
